+++ NEWS.md
@@ -10,6 +10,7 @@ border-radius: 128px;
 
 - Features
     - Browser filters now show selected filters more clearly in the menus.
+    - Fix filtering on credits when clicking on metadata chips.
     - OPDS v2 credits and subjects now link to views filtered on the tag.
 
 ## v1.9.19